working
git-svn-id: svn+ssh://beastie.cs.ua.edu/var/www/svn/scam@361 38ebec9f-8de8-411d-bf7c-b5c56b76c706

+++ prim.c
@@ -1862,7 +1862,7 @@ isGensym(int args)
 
     if (type(id) != SYMBOL) return FalseSymbol;
 
-    return scamBoolean(isdigit(*SymbolTable[ival(id)]));
+    return scamBoolean(isdigit((int)(*SymbolTable[ival(id)])));
     }
 
 /*******ports*********************************/
